[maven-release-plugin] copy for tag xchange-stream-parent-4.3.11

+++ xchange-bitstamp/src/main/java/info/bitrich/xchangestream/bitstamp/dto/BitstampWebSocketTransaction.java
@@ -1,12 +1,14 @@
 package info.bitrich.xchangestream.bitstamp.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+
 import org.knowm.xchange.bitstamp.dto.marketdata.BitstampTransaction;
 
 import java.math.BigDecimal;
 
 public class BitstampWebSocketTransaction extends BitstampTransaction {
-    public BitstampWebSocketTransaction(@JsonProperty("datetime") long date, @JsonProperty("id") int tid, @JsonProperty("price") BigDecimal price, @JsonProperty("amount") BigDecimal amount, @JsonProperty("order_type") int type) {
+    public BitstampWebSocketTransaction(@JsonProperty("datetime") long date, @JsonProperty("id") long tid, @JsonProperty("price") BigDecimal price,
+                                        @JsonProperty("amount") BigDecimal amount, @JsonProperty("order_type") int type) {
         super(date, tid, price, amount, type);
     }
 }
